rename FileService to extract an interface to be consistent
